pyproject.toml
[project]
name = "turtlequant"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.0",
    "numpy>=2.0.0",
    "scipy>=1.13.0",
    "requests>=2.32.5",
    "python-dateutil>=2.9.0",
    "rich>=13.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/turtlequant"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
